fix: key partition
